Finish ...

+++ sharexe-client/src/actions/auth.action.js
@@ -1,5 +1,9 @@
 import * as actionTypes from '../constants/actionTypes';
 import * as authServices from '../services/auth.service';
+import * as socketEvents from '../constants/socketEvents'
+import * as messageServices from '../services/message.service';
+
+import Socket from '../socket';
 
 export function logInUser(username, password) {
     return async function (dispatch) {
@@ -8,6 +12,8 @@ export function logInUser(username, password) {
             localStorage.setItem('token', token);
             const { data: user } = await authServices.getMe();
             dispatch({ type: actionTypes.LOGIN_USER_SUCCESSFUL, payload: { user } });
+            const { data: rooms } = await messageServices.getRecentContacts();
+            Socket.getInstance().emit(socketEvents.THIS_USER_GOES_ONLINE, { rooms: rooms.map(room => room.roomId), userId: user.id });
         } catch (e) {
             console.log(e);
             dispatch({ type: actionTypes.LOGIN_USER_FAILED });
@@ -23,6 +29,8 @@ export function getMe() {
             try {
                 const { data: user } = await authServices.getMe(token);
                 dispatch({ type: actionTypes.LOGIN_USER_SUCCESSFUL, payload: { user } });
+                const { data: rooms } = await messageServices.getRecentContacts();
+                Socket.getInstance().emit(socketEvents.THIS_USER_GOES_ONLINE, { rooms: rooms.map(room => room.roomId), userId: user.id });
             } catch (e) {
                 console.log(e);
                 dispatch({ type: actionTypes.LOGIN_USER_FAILED });
@@ -32,7 +40,10 @@ export function getMe() {
     }
 }
 
-export function logOutUser() {
-    localStorage.clear();
+export function logOutUser(userId, isClosed) {
+    if (!isClosed) {
+        localStorage.clear();
+    }
+    Socket.getInstance().emit(socketEvents.THIS_USER_GOES_OFFLINE, { userId });
     return { type: actionTypes.LOGOUT_USER };
 }
